move common3
